Created a base class for the ARK Loop Function (ALF). Rewrote the two example using ALF based on the created class. Fix grammar, coding style and bugs.
