+++ src/examples/loop_functions/ARK_loop_functions/clustering/CMakeLists.txt
@@ -1,6 +1,6 @@
 include_directories(argos3/plugins/robot/kilobot/control_interface argos3/plugins/robot/kilobot/simulator)
 
-add_library(ALF_clustering_loop_function MODULE ALF.h ALF.cpp)
+add_library(ALF_clustering_loop_function MODULE Clustring_ALF.h Clustring_ALF.cpp)
 
 target_link_libraries(ALF_clustering_loop_function
   argos3core_simulator
