+++ src/examples/loop_functions/ARK_loop_functions/demoC/ALF_qt_user_functions.cpp
@@ -1,5 +1,5 @@
 #include "ALF_qt_user_functions.h"
-#include "ALF.h"
+#include "demoC_ALF.h"
 
 using namespace argos;
